# openapi.yaml
openapi: "3.1.0"
info:
  title: CodeLens — AI Code Review
  description: >
    AI-powered code review GPT Action. Submit a GitHub PR URL or code snippet
    and receive structured review findings with severity ratings, security flags,
    performance suggestions, and refactoring tips.
  version: "1.0.0"
  contact:
    name: CodeLens Team
    url: https://github.com/codelens
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://codelens-api.railway.app
    description: Production API
  - url: http://localhost:8000
    description: Local development

paths:
  /review-pr:
    post:
      operationId: reviewPullRequest
      summary: Review a GitHub Pull Request
      description: >
        Submit a GitHub PR URL for AI-powered code review. Returns structured
        findings with severity, category, line references, and actionable suggestions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewPRRequest"
      responses:
        "200":
          description: Successful review with structured findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResult"
        "400":
          description: Invalid PR URL format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: PR not found or repository is private
        "429":
          description: Rate limit exceeded (10 reviews/hour)

  /review-snippet:
    post:
      operationId: reviewCodeSnippet
      summary: Review a code snippet
      description: >
        Submit a code snippet for AI-powered review. Optionally specify the
        programming language for more accurate analysis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewSnippetRequest"
      responses:
        "200":
          description: Successful review with findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResult"
        "400":
          description: Empty or oversized code snippet
        "429":
          description: Rate limit exceeded

  /health:
    get:
      operationId: healthCheck
      summary: Health check
      description: Returns service health status including database and Redis connectivity.
      responses:
        "200":
          description: Service health status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

components:
  schemas:
    ReviewPRRequest:
      type: object
      required:
        - repo_url
      properties:
        repo_url:
          type: string
          description: "Full GitHub PR URL, e.g. https://github.com/owner/repo/pull/123"
          example: "https://github.com/facebook/react/pull/28000"

    ReviewSnippetRequest:
      type: object
      required:
        - code_snippet
      properties:
        code_snippet:
          type: string
          description: The code to review
          maxLength: 50000
        language:
          type: string
          description: Programming language (auto-detected if not specified)
          default: auto
          example: python
        context:
          type: string
          description: Optional context about the code's purpose
          nullable: true

    ReviewResult:
      type: object
      properties:
        review_id:
          type: string
          description: Unique review identifier
        status:
          type: string
          enum: [pending, completed, error]
        summary:
          type: string
          description: High-level summary of the review
        overall_quality:
          type: string
          enum: [excellent, good, needs_improvement, poor]
        severity_counts:
          $ref: "#/components/schemas/SeverityCounts"
        top_priority_fixes:
          type: array
          items:
            type: string
          maxItems: 5
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReviewItem"
        files_reviewed:
          type: integer
        cached:
          type: boolean
        review_time_ms:
          type: integer

    SeverityCounts:
      type: object
      properties:
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
        info:
          type: integer

    ReviewItem:
      type: object
      properties:
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        category:
          type: string
          enum: [security, performance, bug, style, refactoring, best_practice, documentation, error_handling, testing]
        file_path:
          type: string
          nullable: true
        line_range:
          type: string
          nullable: true
        title:
          type: string
        suggestion:
          type: string
        explanation:
          type: string
        code_before:
          type: string
          nullable: true
        code_after:
          type: string
          nullable: true

    HealthResponse:
      type: object
      properties:
        status:
          type: string
        version:
          type: string
        database:
          type: string
        redis:
          type: string
        github_api:
          type: string
        uptime_seconds:
          type: number

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        trace_id:
          type: string
          nullable: true